refactor: update pkgs and lint code
