+++ frontend/src/hooks/useStylesHome.ts
